Adding three new public methods to the transport.Header struct: WithRaw(k, v string), AllItems() and RawItems()
The goal is to allow outbound middleware to inject headers that are not mutated by the underlying transport. This change
proposes using a new RawItems bag rather than modifying the existing behavior of canonicalized Items / OriginalItems.

After reviewing the transport implementations, the OriginalItems are already used in some contexts and could be replaced
with the RawHeaders if desired. This change also modifies the behavior of Get(), Del() to support RawItems. If there is
a key collision between Items and RawItems, we always return the Items value to avoid regressing existing behavior.

AllItems() is exposed to give a performant way for transports to read all headers without having to do a merge between Items and RawItems.
AllItems is maintained internally to provide the same behavior as Get, favoring the Items value if there is a collision.